fixed build errors
